pyproject.toml
[project]
name = "project-aug"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datetime>=5.5",
    "fastapi>=0.116.1",
    "grad-cam>=1.5.5",
    "jupyter>=1.1.1",
    "koreanize-matplotlib>=0.1.1",
    "matplotlib>=3.10.5",
    "numpy>=2.2.6",
    "optuna>=4.5.0",
    "pandas>=2.3.1",
    "pathlib>=1.0.1",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "streamlit>=1.48.1",
    "surprise>=0.1",
    "tensor>=0.3.6",
    "tensorboard>=2.20.0",
    "tensorboardx>=2.6.4",
    "torchmetrics>=1.8.1",
    "torchsummary>=1.5.1",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "transformers>=4.55.3",
    "uvicorn>=0.35.0",
    "wandb>=0.21.1",
    "xgboost>=3.0.4",
]
